rename lowering from mir to lir module
